Support absolute sizes in ems
Users who want to use semi-absolute sizes but specify the image size in
ems (eg to maintain a good vertical rhythm) can now do so.

In order to
specify the image size in ems, the output html adds an inline style to
the img tag with the width and height, and treats the wrapping figure
the same way as done with absolute-sized images (so the wrapping figure
element can shrink to fit the img, with the correct css).